Address Karthik's comment
