+++ data/src/main/java/com/linkedin/data/schema/annotation/DataSchemaRichContextTraverser.java
@@ -173,25 +173,6 @@ public class DataSchemaRichContextTraverser
   }
 
   /**
-   * Enum to tell how the current schema is linked from its parentSchema as a child schema
-   */
-  enum CurrentSchemaEntryMode
-  {
-    // child schema is for a record's field
-    FIELD,
-    // child schema is the key field of map
-    MAP_KEY,
-    // child schema is the value field of map
-    MAP_VALUE,
-    // child schema is the item of array
-    ARRAY_VALUE,
-    // child schema is a member of union
-    UNION_MEMBER,
-    // child schema is referred from a typeref schema
-    TYPEREF_REF
-  }
-
-  /**
    * Returns true if the dataSchema is a leaf node.
    *
    * a leaf DataSchema is a schema that doesn't have other types of DataSchema linked from it.
@@ -208,165 +189,5 @@ public class DataSchemaRichContextTraverser
     return (dataSchema instanceof PrimitiveDataSchema)
            || (dataSchema.getType() == DataSchema.Type.FIXED)
            || (dataSchema.getType() == DataSchema.Type.ENUM);
-  }
-
-  private static class TraverserContextImpl implements TraverserContext
-  {
-    private Boolean _shouldContinue = null;
-    private DataSchema _parentSchema;
-    private DataSchema _currentSchema;
-    private DataSchema _originalTopLevelSchema;
-    private ArrayDeque<String> _traversePath = new ArrayDeque<>();
-    private ArrayDeque<String> _schemaPathSpec = new ArrayDeque<>();
-    private RecordDataSchema.Field _enclosingField;
-    private UnionDataSchema.Member _enclosingUnionMember;
-    private CurrentSchemaEntryMode _currentSchemaEntryMode;
-    private SchemaVisitor.VisitorContext _visitorContext;
-
-    @Override
-    public SchemaVisitor.VisitorContext getVisitorContext()
-    {
-      return _visitorContext;
-    }
-
-    /**
-     * Generate a new {@link TraverserContext} for next recursion in {@link DataSchemaRichContextTraverser#doRecursiveTraversal(TraverserContextImpl)}
-     *
-     * @param nextTraversePathComponent pathComponent of the traverse path of the next dataSchema to be traversed
-     * @param nextSchemaPathSpecComponent pathComponent of the schema path of the next dataSchema to be traversed
-     * @param nextSchema the next dataSchema to be traversed
-     * @param nextSchemaEntryMode how next dataSchema is linked from current dataSchema.
-     * @return a new {@link TraverserContext} generated for next recursion
-     */
-    private TraverserContextImpl getNextContext(String nextTraversePathComponent, String nextSchemaPathSpecComponent,
-                                    DataSchema nextSchema, CurrentSchemaEntryMode nextSchemaEntryMode)
-    {
-      TraverserContextImpl nextContext = new TraverserContextImpl();
-      nextContext.setOriginalTopLevelSchema(this.getTopLevelSchema());
-      nextContext.setParentSchema(this.getCurrentSchema());
-      nextContext.setSchemaPathSpec(new ArrayDeque<>(this.getSchemaPathSpec()));
-      nextContext.setVisitorContext(this.getVisitorContext());
-      nextContext.setEnclosingField(this.getEnclosingField());
-      nextContext.setEnclosingUnionMember(this.getEnclosingUnionMember());
-
-      // Need to make a copy so if it is modified in next recursion
-      // it won't affect this recursion
-      nextContext.setTraversePath(new ArrayDeque<>(this.getTraversePath()));
-      nextContext.getTraversePath().add(nextTraversePathComponent);
-      // Same as traversePath, we need to make a copy.
-      nextContext.setSchemaPathSpec(new ArrayDeque<>(this.getSchemaPathSpec()));
-      // SchemaPathSpecComponent could be null if nextSchema is a TypeRefDataSchema
-      if (nextSchemaPathSpecComponent != null)
-      {
-        nextContext.getSchemaPathSpec().add(nextSchemaPathSpecComponent);
-      }
-      nextContext.setCurrentSchema(nextSchema);
-      nextContext.setCurrentSchemaEntryMode(nextSchemaEntryMode);
-      return nextContext;
-    }
-
-    private void setOriginalTopLevelSchema(DataSchema originalTopLevelSchema)
-    {
-      _originalTopLevelSchema = originalTopLevelSchema;
-    }
-
-    public Boolean shouldContinue()
-    {
-      return _shouldContinue;
-    }
-
-    @Override
-    public void setShouldContinue(Boolean shouldContinue)
-    {
-      this._shouldContinue = shouldContinue;
-    }
-
-    public void setVisitorContext(SchemaVisitor.VisitorContext visitorContext)
-    {
-      _visitorContext = visitorContext;
-    }
-
-    @Override
-    public DataSchema getTopLevelSchema()
-    {
-      return _originalTopLevelSchema;
-    }
-
-    @Override
-    public ArrayDeque<String> getSchemaPathSpec()
-    {
-      return _schemaPathSpec;
-    }
-
-    private void setSchemaPathSpec(ArrayDeque<String> schemaPathSpec)
-    {
-      _schemaPathSpec = schemaPathSpec;
-    }
-
-    @Override
-    public DataSchema getCurrentSchema()
-    {
-      return _currentSchema;
-    }
-
-    private void setCurrentSchema(DataSchema currentSchema)
-    {
-      _currentSchema = currentSchema;
-    }
-
-    @Override
-    public ArrayDeque<String> getTraversePath()
-    {
-      return _traversePath;
-    }
-
-    private void setTraversePath(ArrayDeque<String> traversePath)
-    {
-      this._traversePath = traversePath;
-    }
-
-    @Override
-    public DataSchema getParentSchema()
-    {
-      return _parentSchema;
-    }
-
-    private void setParentSchema(DataSchema parentSchema)
-    {
-      _parentSchema = parentSchema;
-    }
-
-    @Override
-    public RecordDataSchema.Field getEnclosingField()
-    {
-      return _enclosingField;
-    }
-
-    private void setEnclosingField(RecordDataSchema.Field enclosingField)
-    {
-      _enclosingField = enclosingField;
-    }
-
-    @Override
-    public UnionDataSchema.Member getEnclosingUnionMember()
-    {
-      return _enclosingUnionMember;
-    }
-
-    private void setEnclosingUnionMember(UnionDataSchema.Member enclosingUnionMember)
-    {
-      _enclosingUnionMember = enclosingUnionMember;
-    }
-
-    @Override
-    public CurrentSchemaEntryMode getCurrentSchemaEntryMode()
-    {
-      return _currentSchemaEntryMode;
-    }
-
-    private void setCurrentSchemaEntryMode(CurrentSchemaEntryMode currentSchemaEntryMode)
-    {
-      _currentSchemaEntryMode = currentSchemaEntryMode;
-    }
   }
 }
